Addressing some concerns from Zidane.
